Rename withdraw-families tools to families_withdrawal_{genotypes,phenotypes}
Rename the two CLI tools and their shared helper for a consistent
families_withdrawal_* scheme:

- genotypes_withdraw_families  -> families_withdrawal_genotypes
- phenotypes_withdraw_families -> families_withdrawal_phenotypes
- withdraw_families_common     -> families_withdrawal_common  (shared helper)

Renames the console_scripts entry points, the module files (git mv,
history preserved), the test files, the wrong-type redirect map, and the
docstrings. Also corrects a stale docstring that still described the
genotype tool as rewriting family-variant files (it is pedigree-only).
