+++ core/pyproject.toml
@@ -92,8 +92,8 @@ denovo2vcf = "gpf.tools.denovo2vcf:main"
 dae2vcf = "gpf.tools.dae2vcf:main"
 vcf2tsv = "gpf.tools.vcf2tsv:main"
 verify_parquet = "gpf.tools.verify_parquet:main"
-genotypes_withdraw_families = "gpf.tools.genotypes_withdraw_families:main"
-phenotypes_withdraw_families = "gpf.tools.phenotypes_withdraw_families:main"
+families_withdrawal_genotypes = "gpf.tools.families_withdrawal_genotypes:main"
+families_withdrawal_phenotypes = "gpf.tools.families_withdrawal_phenotypes:main"
 
 [project.entry-points."gpf.genotype_storage.factories"]
 inmemory = "gpf.inmemory_storage.inmemory_genotype_storage:InmemoryGenotypeStorage"
